test: migrate zero-uploads-complete to bdd
